fix: don't require siteUrl when host and sitemap are set
While the case of siteUrl not beeing present was handled, the code didn't
actually run because the graphql client emits an error when a non present field
is queried. This changes the behaviour so that a query is only executed when
either host or sitemap aren't set.

The change should be non-breaking.